Split C codegen expression tests
